+++ src/config/socialLinks.js
@@ -1,19 +1,19 @@
 const socialLinks = [
     {
         icon: 'facebook',
-        url: 'https://www.facebook.com/tomas.scattini/'
+        url: 'https://www.facebook.com/'
     },
     {
         icon: 'twitter',
-        url: 'https://twitter.com/tomiscattini'
+        url: 'https://twitter.com/'
     },
     {
         icon: 'linkedin',
-        url: 'https://www.linkedin.com/in/tomas-scattini/'
+        url: '/https://www.linkedin.com/'
     },
     {
         icon: 'instagram',
-        url: 'https://instagram.com/tomiscattini/'
+        url: 'https://instagram.com/'
     }
 ];