+++ src/data/rito.js
@@ -1,12 +1,12 @@
 const rito = {
+    id: 'rito',
     title: 'Rito',
     subtitle: 'web/app editorial project',
     overview:
-        'Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.',
-    focus: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.',
-    categories:
-        'Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.',
-    brief: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga. Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.',
+        'Rito is an emergent art magazine to promote new and upcoming artists and review the work of established ones, with the intent to democratize art and enable easier access for everyone both artist and viewers.',
+    focus: 'Experience and Interface Design',
+    categories: 'Website/Mobile ',
+    brief: 'Rito is a collaborative project,  done with Tadeu Carabias in  is a  magazine oriented to a young audience. Our public intended to be an audience who enjoys knowing new and emergent artists, who like to know what happens in the artistic environment, is aware of new trends, and receives recommendations on new artistic creations to continue feeding their artistic interest spirit.',
     research:
         'Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga. Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.',
     visualConcept: